explicit cast

+++ include/stencil-composition/backend_cuda/backend_traits_cuda.hpp
@@ -205,7 +205,7 @@ namespace gridtools {
             const uint_t j = Arg::location_t::n_colors::value *
                              (diff_between_blocks * gridDim.x * processing_element_j() * block_size_j);
             // return field offset (Initial storage offset + Alignment correction value + I offset + J offset)
-            return StorageInfo::get_initial_offset() - CurrentExtent::iminus::value + i + j;
+            return (int)StorageInfo::get_initial_offset() - CurrentExtent::iminus::value + i + j;
         }
 
         /**
